+++ driver-core/src/main/java/com/datastax/driver/core/exceptions/NoHostAvailableException.java
@@ -65,6 +65,22 @@ public class NoHostAvailableException extends DriverException {
     }
 
     private static String makeMessage(Map<InetAddress, String> errors) {
-        return String.format("All host(s) tried for query failed (tried: %s)", errors.keySet());
+        // For small cluster, ship the whole error detail in the error message.
+        // This is helpful when debugging on a localhost/test cluster in particular.
+        if (errors.size() == 0)
+            return String.format("All host(s) tried for query failed (no host was tried)");
+
+        if (errors.size() <= 3) {
+            StringBuilder sb = new StringBuilder();
+            sb.append("All host(s) tried for query failed (tried: ");
+            int n = 0;
+            for (Map.Entry<InetAddress, String> entry : errors.entrySet())
+            {
+                if (n++ > 0) sb.append(", ");
+                sb.append(entry.getKey()).append(" (").append(entry.getValue()).append(")");
+            }
+            return sb.append(")").toString();
+        }
+        return String.format("All host(s) tried for query failed (tried: %s - use getErrors() for details)", errors.keySet());
     }
 }