+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/BuiltStatement.java
@@ -70,7 +70,7 @@ abstract class BuiltStatement extends Statement {
             return;
 
         for (int i = 0; i < partitionKey.size(); i++) {
-            if (name.equals(partitionKey.get(i).getName()) && !Utils.isFunctionCall(value)) {
+            if (name.equals(partitionKey.get(i).getName()) && Utils.isRawValue(value)) {
                 routingKey[i] = partitionKey.get(i).getType().parse(Utils.toRawString(value));
                 return;
             }