+++ driver-examples/stress/src/main/java/com/datastax/driver/stress/Stress.java
@@ -55,8 +55,9 @@ public class Stress {
         OptionParser parser = new OptionParser();
 
         parser.accepts("?", "Show this help message");
-        parser.accepts("n", "Number of iterations for the query generator").withRequiredArg().ofType(Integer.class).defaultsTo(1000000);
-        parser.accepts("t", "Number of threads to use").withRequiredArg().ofType(Integer.class).defaultsTo(50);
+        parser.accepts("n", "Number of requests to perform (default: unlimited)").withRequiredArg().ofType(Integer.class);
+        parser.accepts("t", "Level of concurrency to use").withRequiredArg().ofType(Integer.class).defaultsTo(50);
+        parser.accepts("async", "Make asynchronous requests instead of blocking ones");
         parser.accepts("csv", "Save metrics into csv instead of displaying on stdout");
         parser.accepts("columns-per-row", "Number of columns per CQL3 row").withRequiredArg().ofType(Integer.class).defaultsTo(5);
         parser.accepts("value-size", "The size in bytes for column values").withRequiredArg().ofType(Integer.class).defaultsTo(34);
@@ -78,6 +79,8 @@ public class Stress {
             System.exit(1);
         }
 
+        QueryGenerator.Builder genBuilder = generators.get(action);
+
         String[] opts = new String[args.length - 1];
         System.arraycopy(args, 1, opts, 0, opts.length);
 
@@ -90,23 +93,30 @@ public class Stress {
             System.exit(1);
         }
 
-        int iterations = (Integer)options.valueOf("n");
-        int threads = (Integer)options.valueOf("t");
+        int requests = options.has("n") ? (Integer)options.valueOf("n") : -1;
+        int concurrency = (Integer)options.valueOf("t");
 
-        QueryGenerator generator = generators.get(action).create(iterations, options);
-
-        boolean async = false;
+        boolean async = options.has("async");
         boolean useCsv = options.has("csv");
 
-        BlockingQueue<QueryGenerator.Request> workQueue = new SynchronousQueue<QueryGenerator.Request>(true);
+        System.out.println("Initializing stress test...");
+        System.out.println("request count: " + (requests == -1 ? "unlimited" : requests));
+        System.out.println("concurrency: " + concurrency);
+        System.out.println("mode: " + (async ? "asynchronous" : "blocking"));
 
         try {
             // Create session to hosts
             Cluster cluster = new Cluster.Builder().addContactPoints(String.valueOf(options.valueOf("ip"))).build();
 
-            //PoolingOptions pools = cluster.getConfiguration().getConnectionsConfiguration().getPoolingOptions();
-            //pools.setCoreConnectionsPerHost(HostDistance.LOCAL, 2);
-            //pools.setMaxConnectionsPerHost(HostDistance.LOCAL, 2);
+            final int maxRequestsPerConnection = 128;
+            int maxConnections = concurrency / maxRequestsPerConnection + 1;
+
+            PoolingOptions pools = cluster.getConfiguration().getPoolingOptions();
+            pools.setMaxSimultaneousRequestsPerConnectionTreshold(HostDistance.LOCAL, concurrency);
+            pools.setCoreConnectionsPerHost(HostDistance.LOCAL, maxConnections);
+            pools.setMaxConnectionsPerHost(HostDistance.LOCAL, maxConnections);
+            pools.setCoreConnectionsPerHost(HostDistance.REMOTE, maxConnections);
+            pools.setMaxConnectionsPerHost(HostDistance.REMOTE, maxConnections);
 
             Session session = cluster.connect();
 
@@ -114,31 +124,25 @@ public class Stress {
             System.out.println(String.format("Connected to cluster '%s' on %s.", metadata.getClusterName(), metadata.getAllHosts()));
 
             System.out.println("Creating schema...");
-            generator.createSchema(session);
+            genBuilder.createSchema(options, session);
 
             Reporter reporter = new Reporter(useCsv);
-            Producer producer = new Producer(generator, workQueue);
 
-            Consumer[] consumers = new Consumer[threads];
-            Consumer.Asynchronous.ResultHandler resultHandler = async ? new Consumer.Asynchronous.ResultHandler() : null;
-            for (int i = 0; i < threads; i++) {
-                consumers[i] = async
-                             ? new Consumer.Asynchronous(session, workQueue, reporter, resultHandler)
-                             : new Consumer(session, workQueue, reporter);
+            Consumer[] consumers = new Consumer[concurrency];
+            for (int i = 0; i < concurrency; i++) {
+                int iterations = (requests  == -1 ? -1 : requests / concurrency);
+                QueryGenerator generator = genBuilder.create(i, iterations, options, session);
+                consumers[i] = async ? new AsynchronousConsumer(session, generator, reporter) :
+                                       new BlockingConsumer(session, generator, reporter);
             }
 
             System.out.println("Starting to stress test...");
-            producer.start();
-            if (resultHandler != null)
-                resultHandler.start();
+
             for (Consumer consumer : consumers)
                 consumer.start();
 
-            producer.join();
             for (Consumer consumer : consumers)
                 consumer.join();
-            if (resultHandler != null)
-                resultHandler.join();
 
             System.out.println("Stress test successful.");
             System.exit(0);