+++ driver-core/CHANGELOG.rst
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+* 1.0.1:
+  - [api] Function call handling in the query builder has been modified in a
+    backward incompatible way. Function calls are not parsed from string values
+    anymore as this wasn't safe. Instead the new 'fcall' method should be used.
+  - [new] null values support in the query builder.
+  - [new] SSL support (requires C* >= 1.2.1).
+  - [new] Allow generating unlogged batch in the query builder.
+  - [improvement] Better error message when no host are available.
+  - [improvement] Improves performance of the stress example application been.
+
 * 1.0.0:
   - [api] The AuthInfoProvider has be (temporarily) removed. Instead, the
     Cluster builder has a new withCredentials() method to provide a username
@@ -39,7 +49,6 @@ CHANGELOG
   - [bug] Fix failover bug in DCAwareRoundRobinPolicy (JAVA-62)
   - [bug] Fix use of bind markers for routing keys in the query builder
     (JAVA-66)
-
 
 * 1.0.0-beta1:
   - initial release